:zap: Refactor creds endpoints into router
