+++ packages/cli/src/Server.ts
@@ -167,9 +167,11 @@ import { DEFAULT_EXECUTIONS_GET_ALL_LIMIT, validateEntity } from './GenericHelpe
 import { ExecutionEntity } from './databases/entities/ExecutionEntity';
 import { SharedWorkflow } from './databases/entities/SharedWorkflow';
 import { RESPONSE_ERROR_MESSAGES } from './constants';
-import { credentialsEndpoints } from './api/namespaces/credentials';
+import { credentialsRouter } from './api/public/credentials.api';
 
 require('body-parser-xml')(bodyParser);
+
+export const externalHooks: IExternalHooksClass = ExternalHooks();
 
 class App {
 	app: express.Application;
@@ -260,7 +262,7 @@ class App {
 		this.sslKey = config.get('ssl_key');
 		this.sslCert = config.get('ssl_cert');
 
-		this.externalHooks = ExternalHooks();
+		this.externalHooks = externalHooks;
 
 		this.presetCredentialsLoaded = false;
 		this.endpointPresetCredentials = config.get('credentials.overwrite.endpoint') as string;
@@ -652,6 +654,8 @@ class App {
 		// User Management
 		// ----------------------------------------
 		await userManagementRouter.addRoutes.apply(this, [ignoredEndpoints, this.restEndpoint]);
+
+		this.app.use(`/${this.restEndpoint}/credentials`, credentialsRouter);
 
 		// ----------------------------------------
 		// Healthcheck
@@ -1556,12 +1560,6 @@ class App {
 				return this.activeWorkflowRunner.getActivationError(workflowId);
 			}),
 		);
-
-		// ----------------------------------------
-		// Credentials
-		// ----------------------------------------
-
-		credentialsEndpoints.apply(this);
 
 		// ----------------------------------------
 		// Credential-Types
